+++ spec/lib/karafka/setup/config_spec.rb
@@ -9,22 +9,20 @@ RSpec.describe Karafka::Setup::Config do
 
   describe '#setup_components' do
     before do
-      Karafka::Setup::Configurators::Base.descendants.each do |descendant_class|
-        config = double
+      config = double
 
-        expect(descendant_class)
-          .to receive(:new)
-          .with(config_class.config)
-          .and_return(config)
-          .at_least(:once)
+      expect(Karafka::Setup::Configurators::WaterDrop)
+        .to receive(:new)
+        .with(config_class.config)
+        .and_return(config)
+        .at_least(:once)
 
-        expect(config)
-          .to receive(:setup)
-          .at_least(:once)
-      end
+      expect(config)
+        .to receive(:setup)
+        .at_least(:once)
     end
 
-    it 'expect to take descendants of BaseConfigurator and run setup on each' do
+    it 'expect to run setup for waterdrop' do
       config_class.send :setup_components
     end
   end