+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - #267 - Styling inconsistency
 - #242 - Support setting the max bytes to fetch per request
 - #247 - Support SCRAM once released
+- Lower ActiveSupport dependency
+- Remove configurators in favor of the after_boot block configurator
 
 ## 1.1.2
 - #256 - Default kafka.seed_brokers configuration is created in invalid format